Extract transaction hash helpers
